backend/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.0,<4.0
psycopg2-binary>=2.8,<3.0
python-socketio>=5.0,<6.0
redis>=3.5.3,<4
Markdown
whitenoise
